% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monotonicity-contraints.R
\name{add_one_point}
\alias{add_one_point}
\title{Add one point in order to satisfy monotonicity constraint}
\usage{
add_one_point(x0, x1, y0, y1, s0, s1, a0, a1)
}
\arguments{
\item{x0}{Lower bound of the interval.}

\item{x1}{Upper bound of the interval.}

\item{y0}{Value of the spline at \code{x0}.}

\item{y1}{Value of the spline at \code{x1}.}

\item{s0}{Slope of the spline at \code{x0}.}

\item{s1}{Slope of the spline at \code{x1}.}

\item{a0}{Second derivative of the spline at \code{x0}.}

\item{a1}{Second derivative of the spline at \code{x1}.}
}
\value{
A list with four components:
\describe{
    \item{\code{x_new}}{The position of the new point.}
    \item{\code{y_new}}{The value of the spline at the new point.}
    \item{\code{s_new}}{The slope of the spline at the new point.}
    \item{\code{a_new}}{The second derivative of the spline at the new point.}
}
}
\description{
Add an additional point to a spline to ensure monotonicity
of the quantile function.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}
