% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monotonicity-contraints.R
\name{add_two_points}
\alias{add_two_points}
\title{Add two points in order to satisfy monotonicity constraint}
\usage{
add_two_points(x0, x1, y0, y1, s0, s1, a0, a1)
}
\arguments{
\item{x0}{Lower bound of the interval.}

\item{x1}{Upper bound of the interval.}

\item{y0}{Value of the spline at \code{x0}.}

\item{y1}{Value of the spline at \code{x1}.}

\item{s0}{Slope of the spline at \code{x0}.}

\item{s1}{Slope of the spline at \code{x1}.}

\item{a0}{Second derivative of the spline at \code{x0}.}

\item{a1}{Second derivative of the spline at \code{x1}.}
}
\value{
A list with eight components:
\describe{
    \item{\code{x_new1}}{The position of the first new point.}
    \item{\code{y_new1}}{The value of the spline at \code{x_new1}.}
    \item{\code{s_new1}}{The slope of the spline at \code{x_new1}.}
    \item{\code{a_new1}}{The second derivative of the spline at \code{x_new1}.}
    \item{\code{x_new2}}{The position of the second new point.}
    \item{\code{y_new2}}{The value of the spline at \code{x_new2}.}
    \item{\code{s_new2}}{The slope of the spline at \code{x_new2}.}
    \item{\code{a_new2}}{The second derivative of the spline at \code{x_new2}.}
}
}
\description{
Add two additional points to a spline to ensure monotonicity
of the quantile function.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}
