% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-up.R
\name{addup_dist}
\alias{addup_dist}
\title{Add up two components of income or wealth}
\usage{
addup_dist(dist1, dist2, theta)
}
\arguments{
\item{dist1}{The distribution of the first component.}

\item{dist2}{The distribution of the second component.}

\item{theta}{The parameter of the Gumbel copula.}
}
\value{
An object of class \code{gpinter_dist_addup}
}
\description{
Add up two components with their own distributions (say, labor
and capital income), assuming the dependence between both components is
characterized by a Gumbel copula.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

