% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubic-spline.R
\name{basis_functions_cubic_spline}
\alias{basis_functions_cubic_spline}
\alias{g00}
\alias{g01}
\alias{g10}
\alias{g11}
\alias{g20}
\alias{g21}
\alias{g00d1}
\alias{g01d1}
\alias{g10d1}
\alias{g11d1}
\alias{g20d1}
\alias{g21d1}
\title{Basis functions for the cubic spline}
\usage{
g00(x)
}
\arguments{
\item{x}{A value in [0, 1], at which the functions are evaluated.}
}
\value{
The value of the function.
}
\description{
Basis functions and their derivatives for the cubic Hermite
spline.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}
