% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-functions.R
\name{bracket_average}
\alias{bracket_average}
\title{Bracket average for generalized Pareto interpolation}
\usage{
bracket_average(dist, p1, p2, ...)
}
\arguments{
\item{dist}{An object of class \code{gpinter_dist_orig}, \code{gpinter_dist_indiv},
\code{gpinter_dist_addup} or \code{gpinter_dist_merge}.}

\item{p1}{A vector of probabilities in [0, 1].}

\item{p2}{A vector of probabilities in [0, 1], same length as \code{p1}, with \code{p2 > p1}.}

\item{...}{Ignored.}
}
\value{
The value of the average in the bracket the bracket [100*p1%, 100*p2%].
}
\description{
Compute the average between the \code{p1}-th and \code{p2}-th
quantiles for a distribution estimated via generalized Pareto interpolation.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

