% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-functions.R
\name{bracket_share}
\alias{bracket_share}
\title{Bracket share for generalized Pareto interpolation}
\usage{
bracket_share(dist, p1, p2, ...)
}
\arguments{
\item{dist}{An object of class \code{gpinter_dist_orig}, \code{gpinter_dist_indiv},
\code{gpinter_dist_addup} or \code{gpinter_dist_merge}.}

\item{p1}{A vector of probabilities in [0, 1].}

\item{p2}{A vector of probabilities in [0, 1].}

\item{...}{Ignored.}
}
\value{
The value of the share of the bracket [100*p1%, 100*p2%].
}
\description{
Compute the value of the share of the bracket [\code{p1}, \code{p2}]
for a distribution estimated via generalized Pareto interpolation.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

