% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivative-estimation.R
\name{central_derivative}
\alias{central_derivative}
\title{Estimate the value of a derivative on an irregular grid using the
central difference method}
\usage{
central_derivative(x0, x1, x2, y0, y1, y2)
}
\arguments{
\item{x0}{The left point of the grid.}

\item{x1}{The central point of the grid.}

\item{x2}{The right point if the grid.}

\item{y0}{The value of the function at \code{x0}.}

\item{y1}{The value of the function at \code{x1}.}

\item{y2}{The value of the function at \code{x2}.}
}
\value{
The value of the derivative at \code{x1}.
}
\description{
Estimate the value of the derivative of a function based on
three irregularly-spaced points. The method is based on a second-order
Taylor expansion of the function around \code{x1}.
}
\author{
Thomas Blanchet
}

