% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline-error.R
\name{central_peano_kernel}
\alias{central_peano_kernel}
\title{Peano kernel for the interpolation method in central intervals}
\usage{
central_peano_kernel(x, t, x0, x1, x2, x3)
}
\arguments{
\item{x}{A value between \code{x1} and \code{x2}.}

\item{t}{A value between \code{x1} and \code{x2} (integration variable).}

\item{x0}{The first point of the grid.}

\item{x1}{The second point of the grid.}

\item{x2}{The third point of the grid.}

\item{x3}{The fourth point of the grid.}
}
\value{
The value of the Peano kernel.
}
\description{
The Peano kernel correspond the spline interpolation error of
the function \eqn{max{(x - t)^3, 0}}, which is the term inside the integral
of the Taylor remainder.
}
\author{
Thomas Blanchet
}

