% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline-error-degree3.R
\name{central_spline_error_degree3}
\alias{central_spline_error_degree3}
\title{Spline error for the central intervals with known derivatives}
\usage{
central_spline_error_degree3(x, x0, x1, x2, x3, upperbound)
}
\arguments{
\item{x}{A value between \code{x1} and \code{x2}.}

\item{x0}{The first point of the grid.}

\item{x1}{The second point of the grid.}

\item{x2}{The third point of the grid.}

\item{x3}{The fourth point of the grid.}

\item{upperbound}{A function that majorates the fourth derivative.}
}
\value{
The value of the bound (up to a multiplicative constant) at x.
}
\description{
Gives the bound on the spline error for central intervals.
}
\author{
Thomas Blanchet
}

