% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quintic-spline.R
\name{clamped1_quintic_spline_noderiv}
\alias{clamped1_quintic_spline_noderiv}
\title{Estimate natural quintic spline with unknown first derivatives
with second derivative clamped at the last point}
\usage{
clamped1_quintic_spline_noderiv(xk, yk, an)
}
\arguments{
\item{xk}{A vector of interpolation points.}

\item{yk}{A vector of values at each interpolation point.}

\item{an}{The value of the second derivative at the last point.}
}
\value{
A list with two components:
\itemize{
    \item{\code{sk}} A vector of first derivatives at each point.
    \item{\code{ak}} A vector of second derivatives at each point.
}
}
\description{
Estimate a natural quintic spline from known values of the
function and the value of the first derivative at the first knot. That is,
it estimates the first and second derivatives at each point by ensuring
that the third and fourth derivatives are continuous at each knot.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

