% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubic-spline.R
\name{clamped_cubic_spline}
\alias{clamped_cubic_spline}
\title{Estimate a natural cubic spline with first derivative clamped at the last point}
\usage{
clamped_cubic_spline(xk, yk, sn)
}
\arguments{
\item{xk}{A vector of interpolation points.}

\item{yk}{A vector of values at each interpolation point.}

\item{sn}{Derivative at the last knot.}
}
\value{
The vector \code{sk} of first derivatives at each knot.
}
\description{
Estimate a natural cubic spline from known values of the
function: that is, estimate the first derivative of the spline at each knot
to ensure a continuous second derivative, a zero second derivative
at the first point, and a given first derivative at the last point.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

