% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quintic-spline.R
\name{clamped_quintic_spline}
\alias{clamped_quintic_spline}
\title{Estimate a quintic spline known first derivatives with second
derivative clamped at the last point}
\usage{
clamped_quintic_spline(xk, yk, sk, an)
}
\arguments{
\item{xk}{A vector of interpolation points.}

\item{yk}{A vector of values at each interpolation point.}

\item{sk}{A vector of slopes at each interpolation point.}

\item{an}{Second derivative at the last knot.}
}
\value{
The vector \code{ak} of second derivatives at each knot.
}
\description{
Estimate a natural quintic spline from known values of the
function and its derivatives: that is, estimate the second derivative of
the spline at each knot to ensure a continuous third derivative and
a zero third derivative at the extremities.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}
