% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-input.R
\name{clean_input_shares}
\alias{clean_input_shares}
\title{Clean (and check) the inputs for \code{shares_fit}}
\usage{
clean_input_shares(p, average, bracketshare = NULL, topshare = NULL,
  bracketavg = NULL, topavg = NULL, first_threshold = NULL,
  bottom_model = NULL, lower_bound = 0, top_model = NULL)
}
\arguments{
\item{p}{A vector of values in [0, 1].}

\item{average}{The average over the entire distribution.}

\item{bracketshare}{The corresponding bracket share.}

\item{topshare}{The corresponding top share.}

\item{bracketavg}{The corresponding bracket average.}

\item{topavg}{The corresponding top average.}

\item{first_threshold}{The value of the first threshold. If \code{NULL}, it
is estimated from the data. Default is \code{NULL}.}

\item{bottom_model}{Which model to use at the bottom of the distribution?
Only relevant if \code{min(p) > 0}. Either \code{"gpd"} for the generalized
Pareto distribution, or \code{"hist"} for histogram density. Default is
\code{"hist"} if \code{min(threshold) > 0}, and \code{"gpd"} otherwise.}

\item{lower_bound}{Lower bound of the distribution. Only relevant if
\code{min(p) > 0}. Default is \code{0}.}

\item{top_model}{Which model to use at the top of the distribution?
Either \code{"gpd"} for generalized Pareto distribution, or \code{"pareto"}
for standard Pareto distribution. Default is \code{"gpd"}.}
}
\description{
Check the validity and consistency of the input arguments
of \code{shares_fit}.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}
