% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-input.R
\name{clean_input_tabulation}
\alias{clean_input_tabulation}
\title{Clean (and check) the inputs for \code{tabulation_fit}}
\usage{
clean_input_tabulation(p, threshold, average = NULL, bracketshare = NULL,
  topshare = NULL, bracketavg = NULL, topavg = NULL, invpareto = NULL,
  bottom_model = NULL, lower_bound = 0)
}
\arguments{
\item{p}{A vector of values in [0, 1].}

\item{threshold}{The quantiles corresponding to \code{p}.}

\item{average}{The average over the entire distribution.}

\item{bracketshare}{The corresponding bracket share.}

\item{topshare}{The corresponding top share.}

\item{bracketavg}{The corresponding bracket average.}

\item{topavg}{The corresponding top average.}

\item{invpareto}{The inverted Pareto coefficient.}

\item{bottom_model}{Which model to use at the bottom of the distribution?
Only relevant if \code{min(p) > 0}. Either \code{"gpd"} for the generalized
Pareto distribution, \code{"hist"} for histogram density, or \code{"dirac"}.
Default is \code{"hist"} if \code{min(threshold) > 0}, \code{"dirac"} if
\code{min(threshold) == 0} and \code{"gpd"} otherwise.}

\item{lower_bound}{Lower bound of the distribution. Only relevant if
\code{min(p) > 0}. Default is \code{0}.}
}
\value{
A list with the following components: \itemize{
    \item
}
}
\description{
Check the validity and consistency of the input arguments
of \code{tabulation_fit}.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}
