% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-input.R
\name{clean_input_thresholds}
\alias{clean_input_thresholds}
\title{Clean (and check) the inputs for \code{threshold_fit}}
\usage{
clean_input_thresholds(p, threshold, average = NULL, last_bracketavg = NULL,
  last_invpareto = NULL, bottom_model = NULL, lower_bound = 0,
  binf = NULL)
}
\arguments{
\item{p}{A vector of values in [0, 1].}

\item{threshold}{The quantiles corresponding to \code{p}.}

\item{average}{The average over the entire distribution. Use \code{NULL} for
unknown. (Default is \code{NULL}.)}

\item{last_bracketavg}{The average in the last bracket. Use \code{NULL} for
unknown. (Default is \code{NULL}.)}

\item{last_invpareto}{The inverted Pareto coefficient at the last threshold.
Use \code{NULL} for unknown. (Default is \code{NULL}.)}

\item{bottom_model}{Which model to use at the bottom of the distribution?
Only relevant if \code{min(p) > 0}. Either \code{"gpd"} for the generalized
Pareto distribution, or \code{"hist"} for histogram density. Default is
\code{"hist"} if \code{min(threshold) > 0}, and \code{"gpd"} otherwise.}

\item{lower_bound}{Lower bound of the distribution. Only relevant if
\code{min(p) > 0}. Default is \code{0}.}

\item{binf}{Asymptotic Pareto coefficient.}
}
\description{
Check the validity and consistency of the input arguments
of \code{threshold_fit}.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}
