% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monotonicity-contraints.R
\name{cns_one_point}
\alias{cns_one_point}
\title{Monotonicity constraints with one additional point}
\usage{
cns_one_point(x0, x1, y0, y1, s0, s1, a0, a1, x_new, y_new, s_new, a_new)
}
\arguments{
\item{x0}{Lower bound of the interval.}

\item{x1}{Upper bound of the interval.}

\item{y0}{Value of the first spline at \code{x0}.}

\item{y1}{Value of the first spline at \code{x1}.}

\item{s0}{Slope of the first spline at \code{x0}.}

\item{s1}{Slope of the first spline at \code{x1}.}

\item{a0}{Second derivative of the first spline at \code{x0}.}

\item{a1}{Second derivative of the first spline at \code{x1}.}

\item{x_new}{The position of the additional point.}

\item{y_new}{The value of the second spline at the additional point.}

\item{s_new}{The slope of the second spline at the additional point.}

\item{a_new}{The second derivative of the second spline at the additional point.}
}
\description{
Set of constraints which must all be negative to ensure
an increasing quantile function, when adding one additional point to a
spline.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

