% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monotonicity-contraints.R
\name{cns_two_points}
\alias{cns_two_points}
\title{Monotonicity constraints with two additional points}
\usage{
cns_two_points(
  x0,
  x1,
  y0,
  y1,
  s0,
  s1,
  a0,
  a1,
  x_new1,
  y_new1,
  s_new1,
  a_new1,
  x_new2,
  y_new2,
  s_new2,
  a_new2
)
}
\arguments{
\item{x0}{Lower bound of the interval.}

\item{x1}{Upper bound of the interval.}

\item{y0}{Value of the first spline at \code{x0}.}

\item{y1}{Value of the first spline at \code{x1}.}

\item{s0}{Slope of the first spline at \code{x0}.}

\item{s1}{Slope of the first spline at \code{x1}.}

\item{a0}{Second derivative of the first spline at \code{x0}.}

\item{a1}{Second derivative of the first spline at \code{x1}.}

\item{x_new1}{The position of the first additional point.}

\item{y_new1}{The value of the second spline at the first additional point.}

\item{s_new1}{The slope of the second spline at the first additional point.}

\item{a_new1}{The second derivative of the second spline at the first additional point.}

\item{x_new2}{The position of the second additional point.}

\item{y_new2}{The value of the second spline at the second additional point.}

\item{s_new2}{The slope of the second spline at the second additional point.}

\item{a_new2}{The second derivative of the second spline at the second additional point.}
}
\description{
Set of constraints which must all be negative to ensure
an increasing quantile function, when adding two additional points to a
spline.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}
