% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individualize.R
\name{couple_share}
\alias{couple_share}
\title{Share of couples above a fractile}
\usage{
couple_share(dist, p, ...)
}
\arguments{
\item{dist}{An object of class \code{gpinter_dist_indiv}.}

\item{p}{A vector of fractiles.}

\item{...}{Ignored.}
}
\value{
A vector with the share of couples above each value of \code{p}.
}
\description{
Give the share of couples above any fractile \code{p} of the
parent distribution of an individualized distribution, interpolated
with PCHIP.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

