% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-decomposition.R
\name{decompose_population}
\alias{decompose_population}
\title{Decompose population origin by percentile in a merged distribution}
\usage{
decompose_population(dist, p, ...)
}
\arguments{
\item{dist}{An object of class \code{gpinter_dist_merge}.}

\item{p}{A numeric vector of fractiles in [0, 1[.}

\item{...}{Ignored.}
}
\value{
A \code{data.frame} with the following columns:
\describe{
    \item{p}{The fractiles defining the brackets.}
    \item{share1}{The population share of the first merged distribution.}
    \item{share2}{The population share of the second merged distribution.}
    \item{...}{The population share of the other merged distribution.}
}
}
\description{
Given an output of \code{merge_dist} and a list of fractiles,
gives the population composition in terms of origin distribution
for each bracket .
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}
