% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv-spline-error-degree3.R
\name{deriv_central_peano_kernel_degree3}
\alias{deriv_central_peano_kernel_degree3}
\title{Derivative of Peano kernel for the interpolation method in central
intervals with known derivatives}
\usage{
deriv_central_peano_kernel_degree3(x, t, x0, x1, x2, x3)
}
\arguments{
\item{x}{A value between \code{x1} and \code{x2}.}

\item{t}{A value between \code{x1} and \code{x2} (integration variable).}

\item{x0}{The first point of the grid.}

\item{x1}{The second point of the grid.}

\item{x2}{The third point of the grid.}

\item{x3}{The fourth point of the grid.}
}
\value{
The value of the derivative if the Peano kernel.
}
\description{
The Peano kernel correspond the spline interpolation error of
the function \eqn{max{(x - t)^3, 0}}, which is the term inside the integral
of the Taylor remainder. This function returns the derivative of the Peano
kernel, which is used to get the interpolation error bound on the derivative
of the interpolation function.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

