% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubic-spline.R
\name{deriv_cubic_spline}
\alias{deriv_cubic_spline}
\title{First derivative of the cubic spline function}
\usage{
deriv_cubic_spline(x, xk, yk, sk)
}
\arguments{
\item{x}{A vector of values between \code{min(xk)} and \code{max(xk)},
at which the function is evaluated.}

\item{xk}{A vector of interpolation points.}

\item{yk}{A vector of values at each interpolation point.}

\item{sk}{A vector of slopes at each interpolation point.}
}
\value{
The values of the first derivative of the interpolation function at
each point \code{x}.
}
\description{
First derivative of the quintic spline interpolation function
over several intervals.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}
