% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv-spline-error-degree2.R
\name{deriv_left_peano_kernel_degree2}
\alias{deriv_left_peano_kernel_degree2}
\title{Derivative of the Peano kernel for the interpolation method in the
left interval with unknown derivatives}
\usage{
deriv_left_peano_kernel_degree2(x, t, x0, x1, x2)
}
\arguments{
\item{x}{A value between \code{x0} and \code{x1}.}

\item{t}{A value between \code{x0} and \code{x1} (integration variable).}

\item{x0}{The first point of the grid.}

\item{x1}{The second point of the grid.}

\item{x2}{The third point of the grid.}
}
\value{
The value of the derivative of the Peano kernel.
}
\description{
The Peano kernel correspond the spline interpolation error of
the function \eqn{max{(x - t), 0}^2}, which is the term inside the integral
of the Taylor remainder. This function estimates the derivative of the
kernel for the interpolation of the left interval, where the derivative
on the right is estimated from the three points to the right.
}
\author{
Thomas Blanchet
}

