% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv-spline-error-degree3.R
\name{deriv_right_peano_kernel_degree3}
\alias{deriv_right_peano_kernel_degree3}
\title{Derivative of the Peano kernel for the interpolation method in the
right interval with known derivatives}
\usage{
deriv_right_peano_kernel_degree3(x, t, x0, x1, x2)
}
\arguments{
\item{x}{A value between \code{x1} and \code{x2}.}

\item{t}{A value between \code{x1} and \code{x2} (integration variable).}

\item{x0}{The first point of the grid.}

\item{x1}{The second point of the grid.}

\item{x2}{The third point of the grid.}
}
\value{
The value of the derivative if the Peano kernel.
}
\description{
The Peano kernel correspond the spline interpolation error of
the function \eqn{max{(x - t)^3, 0}}, which is the term inside the integral
of the Taylor remainder. This function estimates the derivative of the
kernel for the interpolation of the right interval, where the last second
derivative is estimated from the three points to the left.
}
\author{
Thomas Blanchet
}

