% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv-spline-error-degree3.R
\name{deriv_right_spline_error_degree3}
\alias{deriv_right_spline_error_degree3}
\title{Interpolation error of the derivative for the right interval with
known derivatives}
\usage{
deriv_right_spline_error_degree3(x, x0, x1, x2, upperbound)
}
\arguments{
\item{x}{A value between \code{x1} and \code{x2}.}

\item{x0}{The first point of the grid.}

\item{x1}{The second point of the grid.}

\item{x2}{The third point of the grid.}

\item{upperbound}{A function that majorates the fourth derivative.}
}
\value{
The value of the bound at x.
}
\description{
Gives the bound on the interpolation error of the
derivative for the right interval.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

