% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv-spline-error-degree3.R
\name{deriv_spline_error_degree3}
\alias{deriv_spline_error_degree3}
\title{Spline interpolation error of the derivative over the entire domain
with known derivatives}
\usage{
deriv_spline_error_degree3(x, xk, upperbound)
}
\arguments{
\item{x}{A value between \code{min(xk)} and \code{max(xk)}.}

\item{xk}{The grid points for the spline estimation.}

\item{upperbound}{A function that majorates the fourth derivative.}
}
\value{
The value of the bound at x.
}
\description{
Gives the bound on the spline interpolation error
of the derivative for any point of the interpolation domain.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

