% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monotonicity-contraints.R
\name{distance_one_point}
\alias{distance_one_point}
\title{Distance between spline and spline perturbated by one point}
\usage{
distance_one_point(x0, x1, y0, y1, s0, s1, a0, a1, x_new, y_new, s_new, a_new,
  y_ini, s_ini, a_ini)
}
\arguments{
\item{x0}{Lower bound of the interval.}

\item{x1}{Upper bound of the interval.}

\item{y0}{Value of the first spline at \code{x0}.}

\item{y1}{Value of the first spline at \code{x1}.}

\item{s0}{Slope of the first spline at \code{x0}.}

\item{s1}{Slope of the first spline at \code{x1}.}

\item{a0}{Second derivative of the first spline at \code{x0}.}

\item{a1}{Second derivative of the first spline at \code{x1}.}

\item{x_new}{The position of the additional point.}

\item{y_new}{The value of the second spline at the additional point.}

\item{s_new}{The slope of the second spline at the additional point.}

\item{a_new}{The second derivative of the second spline at the additional point.}

\item{y_ini}{The value of the first spline at \code{x_new}.}

\item{s_ini}{The slope of the first spline at \code{x_new}.}

\item{a_ini}{The second derivative of the first spline at \code{x_new}.}
}
\description{
These functions calculate the distance between:
\itemize{
    \item A first spline characterized by two points.
    \item A second set of two splines characterized by the same two points,
    plus an additional point between the first two.
}
They also calculate the derivative of that distance with respect to the
coordinates of the additional point.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

