% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{distance_spline}
\alias{distance_spline}
\alias{distance_spline_deriv_a0new}
\alias{distance_spline_deriv_a1new}
\alias{distance_spline_deriv_s0new}
\alias{distance_spline_deriv_s1new}
\alias{distance_spline_deriv_y0new}
\alias{distance_spline_deriv_y1new}
\title{Distance between two quintic splines}
\usage{
distance_spline(x0, x1, y0, y1, s0, s1, a0, a1, y0new, y1new, s0new, s1new,
  a0new, a1new)
}
\arguments{
\item{x0}{Lower bound of the interval.}

\item{x1}{Upper bound of the interval.}

\item{y0}{Value of the spline at \code{x0} for the first spline.}

\item{y1}{Value of the spline at \code{x1} for the first spline.}

\item{s0}{Slope of the spline at \code{x0} for the first spline.}

\item{s1}{Slope of the spline at \code{x1} for the first spline.}

\item{a0}{Second derivative of the spline at \code{x0} for the first spline.}

\item{a1}{Second derivative of the spline at \code{x1} for the first spline.}

\item{y0new}{Value of the spline at \code{x0} for the first spline.}

\item{y1new}{Value of the spline at \code{x1} for the first spline.}

\item{s0new}{Slope of the spline at \code{x0} for the first spline.}

\item{s1new}{Slope of the spline at \code{x1} for the first spline.}

\item{a0new}{Second derivative of the spline at \code{x0} for the first spline.}

\item{a1new}{Second derivative of the spline at \code{x1} for the first spline.}
}
\value{
The value of the distance or its derivatives.
}
\description{
Calculate the distance between two quintic splines, as
measured by the integral of the square of the difference. Also calculates
the derivatives of that distance with respect to the spline parameters of
the second function.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

