% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monotonicity-contraints.R
\name{distance_two_points}
\alias{distance_two_points}
\title{Distance between spline and spline perturbated by two points}
\usage{
distance_two_points(x0, x1, y0, y1, s0, s1, a0, a1, x_new1, y_new1, s_new1,
  a_new1, y_ini1, s_ini1, a_ini1, x_new2, y_new2, s_new2, a_new2, y_ini2,
  s_ini2, a_ini2)
}
\arguments{
\item{x0}{Lower bound of the interval.}

\item{x1}{Upper bound of the interval.}

\item{y0}{Value of the first spline at \code{x0}.}

\item{y1}{Value of the first spline at \code{x1}.}

\item{s0}{Slope of the first spline at \code{x0}.}

\item{s1}{Slope of the first spline at \code{x1}.}

\item{a0}{Second derivative of the first spline at \code{x0}.}

\item{a1}{Second derivative of the first spline at \code{x1}.}

\item{x_new1}{The position of the first additional point.}

\item{y_new1}{The value of the second spline at the first additional point.}

\item{s_new1}{The slope of the second spline at the first additional point.}

\item{a_new1}{The second derivative of the second spline at the first additional point.}

\item{y_ini1}{The value of the first spline at \code{x_new1}.}

\item{s_ini1}{The slope of the first spline at \code{x_new1}.}

\item{a_ini1}{The second derivative of the first spline at \code{x_new1}.}

\item{x_new2}{The position of the second additional point.}

\item{y_new2}{The value of the second spline at the second additional point.}

\item{s_new2}{The slope of the second spline at the second additional point.}

\item{a_new2}{The second derivative of the second spline at the second additional point.}

\item{y_ini2}{The value of the first spline at \code{x_new2}.}

\item{s_ini2}{The slope of the first spline at \code{x_new2}.}

\item{a_ini2}{The second derivative of the first spline at \code{x_new2}.}
}
\description{
These functions calculate the distance between:
\itemize{
    \item A first spline characterized by two points.
    \item A second set of three splines characterized by the same two points,
    plus two additional points between the first two.
}
They also calculate the derivative of that distance with respect to the
coordinates of the additionals points.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

