% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-functions.R
\name{fitted_cdf}
\alias{fitted_cdf}
\title{Cumulative distribution function for generalized Pareto interpolation}
\usage{
fitted_cdf(dist, x, ...)
}
\arguments{
\item{dist}{An object of class \code{gpinter_dist_orig}, \code{gpinter_dist_indiv},
\code{gpinter_dist_addup} or \code{gpinter_dist_merge}.}

\item{x}{The evaluation point(s).}

\item{...}{Ignored.}
}
\value{
The value of the CDF at \code{x}.
}
\description{
Compute the value of the cumulative distribution function for a
distribution estimated via generalized Pareto interpolation.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}
