% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulation-fit.R
\name{fitted_top_share}
\alias{fitted_top_share}
\title{Top share function}
\usage{
fitted_top_share(p, xk, yk, sk, ak, average, mu_top, sigma_top, xi_top,
  mu_bottom, sigma_bottom, xi_bottom)
}
\arguments{
\item{p}{The function evaluation point.}

\item{xk}{A vector of interpolation points.}

\item{yk}{A vector of values at each interpolation point.}

\item{sk}{A vector of slopes at each interpolation point.}

\item{ak}{A vector of second derivatives at each interpolation point.}

\item{average}{The average over the entire population.}

\item{mu_top}{The location parameter for the generalized Pareto distribution at
the top of the distribution.}

\item{sigma_top}{The scale parameter for the generalized Pareto distribution at
the top of the distribution.}

\item{xi_top}{The shape parameter for the generalized Pareto distribution at
the top of the distribution.}

\item{mu_bottom}{The location parameter for the generalized Pareto distribution at
the bottom of the distribution.}

\item{sigma_bottom}{The scale parameter for the generalized Pareto distribution at
the bottom of the distribution.}

\item{xi_bottom}{The shape parameter for the generalized Pareto distribution at
the bottom of the distribution.}
}
\value{
The value of the share of the top 100*(1 - p)%.
}
\description{
Top share function, combining the quintic spline
at the bottom and the generalized Pareto distribution at the top.
}
\author{
Thomas Blanchet
}

