% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubic-spline.R
\name{gd1}
\alias{gd1}
\title{First derivative of the cubic spline function}
\usage{
gd1(x, x0, x1, y0, y1, s0, s1)
}
\arguments{
\item{x}{A number between \code{x0} and \code{x1}, at which the function is
evaluated.}

\item{x0}{Lower bound of the interval.}

\item{x1}{Upper bound of the interval.}

\item{y0}{Value of the spline at \code{x0}.}

\item{y1}{Value of the spline at \code{x1}.}

\item{s0}{Slope of the spline at \code{x0}.}

\item{s1}{Slope of the spline at \code{x1}.}
}
\value{
The value of first derivative of the interpolation at \code{x}.
}
\description{
First derivative of the cubic spline over an interval
\code{[x0, x1]}.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}
