% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-tabulation.R
\name{generate_tabulation}
\alias{generate_tabulation}
\title{Generate a tabulation of a distribution}
\usage{
generate_tabulation(dist, fractiles, ...)
}
\arguments{
\item{dist}{An object of class \code{gpinter_dist_orig}, \code{gpinter_dist_indiv},
\code{gpinter_dist_addup} or \code{gpinter_dist_merge}.}

\item{fractiles}{A vector of values in [0, 1[.}

\item{...}{Ignored.}
}
\value{
A data frame with a row for each fractile and a column for each
statistic.
}
\description{
Generate a tabulation with shares, averages and thresholds
at different points of the distribution. This methods avoid redundant
computations compared with using the dedicated function for each quantity.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

