% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized-pareto-distribution.R
\name{gpd_bottom_mean}
\alias{gpd_bottom_mean}
\title{Truncated average function of the generalized Pareto distribution in
the bottom of the distribution.}
\usage{
gpd_bottom_mean(p, p1, y1, mu, sigma, xi)
}
\arguments{
\item{p}{A value in [0, p1]. The point at which to evaluate the function.}

\item{p1}{A value in [0, 1]. The point of the distribution at which the
GDP model ends.}

\item{y1}{The value of the interpolation function at x1}

\item{mu}{The location parameter.}

\item{sigma}{The scale parameter.}

\item{xi}{The shape parameter.}
}
\value{
The value of the truncated average at \code{p}.
}
\description{
This function correspond to the truncated average of the
generalized Pareto distribution, plus the value of the truncated average
above.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}
