% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized-pareto-distribution.R
\name{gpd_bottom_parameters}
\alias{gpd_bottom_parameters}
\title{Estimate parameters of the parameters of the generalized Pareto
distribution for the bottom of the distribution}
\usage{
gpd_bottom_parameters(x1, y1, s1, a1, average)
}
\arguments{
\item{x1}{The first interpolation point.}

\item{y1}{The value of the function at the first interpolation point.}

\item{s1}{The value of the first derivative at the first interpolation point.}

\item{a1}{The value of the second derivative at the first interpolation point.}

\item{average}{The average over the entire distribution.}
}
\value{
A list with the three parameters: \code{mu}, \code{sigma} and \code{mu}.
}
\description{
Below the first threshold, we extrapolate the curve using a
generalized Pareto distribution. The three parameters of the distribution
are chosen so as to ensure the continuity and derivability of the quantile
function, and to match the average.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}
