% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized-pareto-distribution.R
\name{gpd_bottom_phi}
\alias{gpd_bottom_phi}
\title{Interpolation function below the first bracket}
\usage{
gpd_bottom_phi(x, p1, y1, mu, sigma, xi)
}
\arguments{
\item{x}{A value lower than x1. The point at which to evaluate the
function.}

\item{p1}{A value in [0, 1]. The point of the distribution at which the
GDP model ends}

\item{y1}{The value of the interpolation function at x1}

\item{mu}{The location parameter.}

\item{sigma}{The scale parameter.}

\item{xi}{The shape parameter.}
}
\value{
The value of the interpolation function at \code{x}.
}
\description{
Interpolation function based on the generalized Pareto
distribution used in the last bracket of the data.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

