% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized-pareto-distribution.R
\name{gpd_bottom_quantile}
\alias{gpd_bottom_quantile}
\title{Quantile function of the generalized Pareto distribution in the bottom
of the distribution.}
\usage{
gpd_bottom_quantile(p, p1, mu, sigma, xi)
}
\arguments{
\item{p}{A value in [0, p1]. The point at which to evaluate the function.}

\item{p1}{A value in [0, 1]. The point of the distribution at which the
GDP model ends}

\item{mu}{The location parameter.}

\item{sigma}{The scale parameter.}

\item{xi}{The shape parameter.}
}
\value{
The value of the quantile at \code{p}.
}
\description{
This function correspond to the quantile function of the
generalized Pareto distribution, but with some rescaling to model the bottom
of the distribution only.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

