% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized-pareto-distribution.R
\name{gpd_top_deriv_phi}
\alias{gpd_top_deriv_phi}
\title{Derivative of the interpolation function in the last bracket}
\usage{
gpd_top_deriv_phi(x, pn, mu, sigma, xi)
}
\arguments{
\item{x}{A value greater than xn. The point at which to evaluate the
function.}

\item{pn}{A value in [0, 1]. The point of the distribution at which the
GDP model starts.}

\item{mu}{The location parameter.}

\item{sigma}{The scale parameter.}

\item{xi}{The shape parameter.}
}
\value{
The value of the derivative of the interpolation function at
\code{x}.
}
\description{
Interpolation function based on the generalized Pareto
distribution used in the last bracket of the data.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

