% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized-pareto-distribution.R
\name{gpd_top_parameters}
\alias{gpd_top_parameters}
\title{Estimate parameters of the parameters of the generalized Pareto
distribution for the top of the distribution}
\usage{
gpd_top_parameters(xn, yn, sn, an)
}
\arguments{
\item{xn}{The last interpolation point.}

\item{yn}{The value of the function at the last interpolation point.}

\item{sn}{The value of the first derivative at the last interpolation point.}

\item{an}{The value of the second derivative at the last interpolation point.}
}
\value{
A list with the three parameters: \code{mu}, \code{sigma} and \code{mu}.
}
\description{
Above the last threshold, we extrapolate the curve using a
generalized Pareto distribution. The three parameters of the distribution
are chosen so as to ensure the continuity and derivability of the quantile
function, and to match the average in the last bracket.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}
