% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-up.R
\name{gumbel_cond_quantile}
\alias{gumbel_cond_quantile}
\title{Conditional quantile function of the Gumbel copula}
\usage{
gumbel_cond_quantile(p, u, theta)
}
\arguments{
\item{p}{A number in [0, 1].}

\item{u}{The value of U.}

\item{theta}{The parameter of the Gumbel copula.}
}
\value{
The p-th quantile of V|U = u.
}
\description{
Assume (U, V) follows a Gumbel copula. This function gives
the quantile function of V given U = u. It is used to simulate the Gumbel
copula.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

