% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quintic-spline.R
\name{h}
\alias{h}
\title{Quintic spline function}
\usage{
h(x, x0, x1, y0, y1, s0, s1, a0, a1)
}
\arguments{
\item{x}{A number between \code{x0} and \code{x1}, at which the spline is
evaluated.}

\item{x0}{Lower bound of the interval.}

\item{x1}{Upper bound of the interval.}

\item{y0}{Value of the spline at \code{x0}.}

\item{y1}{Value of the spline at \code{x1}.}

\item{s0}{Slope of the spline at \code{x0}.}

\item{s1}{Slope of the spline at \code{x1}.}

\item{a0}{Second derivative of the spline at \code{x0}.}

\item{a1}{Second derivative of the spline at \code{x1}.}
}
\value{
The value of the interpolation at \code{x}.
}
\description{
Quintic spline over an interval \code{[x0, x1]}. It is such that:
\itemize{
    \item{\code{h(x0) == y0}}
    \item{\code{h(x1) == y1}}
    \item{\code{h’(x0) == s0}}
    \item{\code{h’(x1) == s1}}
    \item{\code{h’’(x0) == a0}}
    \item{\code{h’’(x1) == a1}}
}
}
\author{
Thomas Blanchet
}

