% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram-interpolation.R
\name{hist_cdf}
\alias{hist_cdf}
\title{CDF for histogram density}
\usage{
hist_cdf(q, p0, q0, f)
}
\arguments{
\item{q}{The values at which to estimate the CDF.}

\item{p0}{The probability lower bound of the interval with constant density.}

\item{q0}{The quantile lower bound of the interval with constant density.}

\item{f}{The value of the constant density.}
}
\description{
Calculates the CDF over an interval where the
distribution has a constant density.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

