% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram-interpolation.R
\name{hist_deriv_phi}
\alias{hist_deriv_phi}
\title{Derivative of Pareto interpolation function for histogram density}
\usage{
hist_deriv_phi(x, p0, p1, q0, q1, m1, f)
}
\arguments{
\item{x}{The values at which to estimate the interpolation function.}

\item{p0}{The probability lower bound of the interval with constant density.}

\item{p1}{The probability upper bound of the interval with constant density.}

\item{q0}{The quantile lower bound of the interval with constant density.}

\item{q1}{The quantile upper bound of the interval with constant density.}

\item{m1}{The value of the Lorenz curve at \code{p1}.}

\item{f}{The value of the constant density.}
}
\description{
Calculates the derivative of the Pareto interpolation function
for a distribution with constant density over an interval.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

