% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram-interpolation.R
\name{hist_interpol}
\alias{hist_interpol}
\title{Histogram interpolation}
\usage{
hist_interpol(p0, p1, q0, q1, average)
}
\arguments{
\item{p0}{A value in [0, 1].}

\item{p1}{A value in [0, 1].}

\item{q0}{The \code{p0}-th quantile.}

\item{q1}{The \code{p1}-th quantile.}

\item{average}{The average value in the bracket.}
}
\value{
A list with the following components:
\describe{
    \item{average}{The average inside the bracket.}
    \item{p0}{A value in [0, 1].}
    \item{p1}{A value in [0, 1].}
    \item{q0}{The \code{p0}-th quantile.}
    \item{q1}{The \code{p1}-th quantile.}
    \item{pstar}{A value in [\code{p0}, \code{p1}].}
    \item{qstar}{A value in [\code{q0}, \code{q1}]: the \code{pstar}-th quantile.}
    \item{f0}{The value of the density in [\code{q0}, \code{qstar}].}
    \item{f1}{The value of the density in [\code{qstar}, \code{q1}].}
}
}
\description{
This function interpolates the distribution over a bracket
with an histogram (that is, piecewise constant density with two values,
and piecewise linear quantile function with two slopes).
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

