% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram-interpolation.R
\name{hist_lorenz}
\alias{hist_lorenz}
\title{Non-normalized Lorenz curve for histogram density}
\usage{
hist_lorenz(p, p0, p1, q0, q1, m1, f)
}
\arguments{
\item{p}{The values at which to estimate the Lorenz curve.}

\item{p0}{The probability lower bound of the interval with constant density.}

\item{p1}{The probability upper bound of the interval with constant density.}

\item{q1}{The quantile lower bound of the interval with constant density.}

\item{m1}{The value of the Lorenz curve at \code{p1}.}

\item{f}{The value of the constant density.}

\item{q1}{The quantile upper bound of the interval with constant density.}
}
\description{
Calculates the non-normalized Lorenz curve (average times the
top share) for a distribution with constant density over an interval.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

