% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation-error.R
\name{interpolation_error_noncons}
\alias{interpolation_deriv_error_noncons}
\alias{interpolation_error_noncons}
\alias{interpolation_value_error_noncons}
\title{Value of the interpolation error for a non constant phi'''}
\usage{
interpolation_value_error_noncons(x, xk, phid3)
}
\arguments{
\item{x}{A vector of points at which to estimate the error.}

\item{xk}{The position of the interpolation points.}

\item{phid3}{The third derivative of the interpolated function.}
}
\value{
The bound for each value of \code{x}.
}
\description{
Give the value of the error for the interpolation and
its derivative, given the third derivative of the
interpolated function.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

