% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-functions.R
\name{invpareto}
\alias{invpareto}
\title{Inverted Pareto coefficients for generalized Pareto interpolation}
\usage{
invpareto(dist, p, ...)
}
\arguments{
\item{dist}{An object of class \code{gpinter_dist_orig}, \code{gpinter_dist_indiv},
\code{gpinter_dist_addup} or \code{gpinter_dist_merge}.}

\item{p}{A vector of probabilities in [0, 1].}

\item{...}{Ignored.}
}
\value{
The value of the inverted Pareto coefficient at \code{p}.
}
\description{
Compute the value of inverted Pareto coefficients for a
distribution estimated via generalized Pareto interpolation.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}
