% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monotonicity-contraints.R
\name{is_increasing}
\alias{is_increasing}
\title{Check of monotonicity constraint}
\usage{
is_increasing(x0, x1, y0, y1, s0, s1, a0, a1, m = 10000)
}
\arguments{
\item{x0}{Lower bound of the interval.}

\item{x1}{Upper bound of the interval.}

\item{y0}{Value of the first spline at \code{x0}.}

\item{y1}{Value of the first spline at \code{x1}.}

\item{s0}{Slope of the first spline at \code{x0}.}

\item{s1}{Slope of the first spline at \code{x1}.}

\item{a0}{Second derivative of the first spline at \code{x0}.}

\item{a1}{Second derivative of the first spline at \code{x1}.}

\item{m}{The number of evaluation points of the polynomial: the bigger,
the slower and the tighter the bounds.}
}
\description{
Check that a spline correspond to an increasing quantile
function, first using the fast condition, then the slower one if necessary.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

