% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monotonicity-contraints.R
\name{is_increasing_fast}
\alias{is_increasing_fast}
\title{Fast check of monotonicity constraint}
\usage{
is_increasing_fast(x0, x1, y0, y1, s0, s1, a0, a1)
}
\arguments{
\item{x0}{Lower bound of the interval.}

\item{x1}{Upper bound of the interval.}

\item{y0}{Value of the first spline at \code{x0}.}

\item{y1}{Value of the first spline at \code{x1}.}

\item{s0}{Slope of the first spline at \code{x0}.}

\item{s1}{Slope of the first spline at \code{x1}.}

\item{a0}{Second derivative of the first spline at \code{x0}.}

\item{a1}{Second derivative of the first spline at \code{x1}.}
}
\description{
Check that a spline correspond to an increasing quantile
function using a fast to compute (but slightly over-restrictive) set of
sufficient conditions.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}
