% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation-error.R
\name{kernel}
\alias{kernel}
\alias{deriv_kernel}
\title{Peano kernel}
\usage{
kernel(x, t)
}
\arguments{
\item{x}{The interpolation variable.}

\item{t}{The integration variable.}
}
\value{
The value of the kernel.
}
\description{
The Peano kernel is the term inside the integral remainder in
the Taylor formula.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}
