% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation-error.R
\name{kernel_interpolation_param}
\alias{kernel_interpolation_param}
\title{Parameters of the spline interpolating the Peano kernel}
\usage{
kernel_interpolation_param(xk, t)
}
\arguments{
\item{xk}{The position of the interpolation points.}

\item{t}{The integration variable.}
}
\value{
A list with the following components:
\itemize{
    \item{\code{xk}} The position of the interpolation points.
    \item{\code{yk}} The value of the spline at the interpolation points.
    \item{\code{sk}} The value of the derivative of the spline at the interpolation points.
    \item{\code{ak}} The value of the second derivative of the spline at the interpolation points.
}
}
\description{
Parameters of the spline interpolating the Peano kernel.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}
