% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation-error.R
\name{kernel_value_error}
\alias{kernel_value_error}
\alias{kernel_deriv_error}
\title{Interpolation error on the kernel}
\usage{
kernel_value_error(xk, x, t)
}
\arguments{
\item{xk}{The position of the interpolation points.}

\item{x}{The point at which to give the error.}

\item{t}{The integration variable in the Peano kernel.}
}
\value{
The value of the interpolation error on the kernel.
}
\description{
Interpolation error on the kernel and its derivative.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}
