% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline-error-degree3.R
\name{left_peano_kernel_degree3}
\alias{left_peano_kernel_degree3}
\title{Peano kernel for the interpolation method with known derivatives
in the left interval}
\usage{
left_peano_kernel_degree3(x, t, x0, x1, x2)
}
\arguments{
\item{x}{A value between \code{x0} and \code{x1}.}

\item{t}{A value between \code{x0} and \code{x1} (integration variable).}

\item{x0}{The first point of the grid.}

\item{x1}{The second point of the grid.}

\item{x2}{The third point of the grid.}
}
\value{
The value of the Peano kernel.
}
\description{
The Peano kernel correspond the spline interpolation error of
the function \eqn{max{(x - t)^3, 0}}, which is the term inside the integral
of the Taylor remainder. This function estimates the kernel for the
interpolation of the left interval, where the first second derivative is
estimated from the three points to the right.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

