% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline-error-degree3.R
\name{left_spline_error_degree3}
\alias{left_spline_error_degree3}
\title{Spline error for the left interval with known derivatives}
\usage{
left_spline_error_degree3(x, x0, x1, x2, upperbound)
}
\arguments{
\item{x}{A value between \code{x1} and \code{x2}.}

\item{x0}{The first point of the grid.}

\item{x1}{The second point of the grid.}

\item{x2}{The third point of the grid.}

\item{upperbound}{A function that majorates the fourth derivative.}
}
\value{
The value of the bound at x.
}
\description{
Gives the bound on the spline error for the left interval.
}
\author{
Thomas Blanchet
}

