% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge_dist}
\alias{merge_dist}
\title{Merge the distribution from different countries}
\usage{
merge_dist(dist, popsize)
}
\arguments{
\item{dist}{A vector of objects of class \code{gpinter_dist}.}

\item{popsize}{A vector of the same length with the population associated
with each distribution.}
}
\value{
An object of class \code{gpinter_dist_merge}.
}
\description{
Takes an arbitrary number of distributions from different
countries, with their respective population sizes, and return an object
that characterizes the combined distribution of all countries.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

