% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{mono_cns}
\alias{mono_cns}
\alias{mono_cns_deriv_a0}
\alias{mono_cns_deriv_a1}
\alias{mono_cns_deriv_s0}
\alias{mono_cns_deriv_s1}
\alias{mono_cns_deriv_y0}
\alias{mono_cns_deriv_y1}
\title{Monotonicity constraint for the quantile function}
\usage{
mono_cns(x0, x1, y0, y1, s0, s1, a0, a1)
}
\arguments{
\item{x0}{Lower bound of the interval.}

\item{x1}{Upper bound of the interval.}

\item{y0}{Value of the spline at \code{x0}.}

\item{y1}{Value of the spline at \code{x1}.}

\item{s0}{Slope of the spline at \code{x0}.}

\item{s1}{Slope of the spline at \code{x1}.}

\item{a0}{Second derivative of the spline at \code{x0}.}

\item{a1}{Second derivative of the spline at \code{x1}.}
}
\value{
The value of the nine constraints or their derivative.
}
\description{
This set of functions provide constraints on the parameters
of the quintic spline that ensure an increasing quantile function. The
function \code{mono_cns} returns nine values that must all be non-negative.
The other are the derivatives of those constraints with respect to each
parameters. Those conditions were derived using the Bernstein polynomial
form of the constraint (see Mathematica workbook for details).
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

