% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quintic-spline.R
\name{natural_quintic_spline}
\alias{natural_quintic_spline}
\title{Estimate natural quintic spline with known first derivatives}
\usage{
natural_quintic_spline(xk, yk, sk)
}
\arguments{
\item{xk}{A vector of interpolation points.}

\item{yk}{A vector of values at each interpolation point.}

\item{sk}{A vector of slopes at each interpolation point.}
}
\value{
The vector \code{ak} of second derivatives at each knot.
}
\description{
Estimate a natural quintic spline from known values of the
function and its derivatives: that is, estimate the second derivative of
the spline at each knot to ensure a continuous third derivative and
a zero third derivative at the extremities.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}
