% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-functions.R
\name{phi}
\alias{phi}
\title{Interpolation function from generalized Pareto interpolation}
\usage{
phi(dist, x, ...)
}
\arguments{
\item{dist}{A \code{gpinter_dist_orig} object, as returned by
\code{tabulation_fit} or \code{share_fit}.}

\item{x}{The function evaluation point(s).}

\item{...}{Ignored.}
}
\value{
The value of the interpolation function at \code{x}.
}
\description{
Full interpolation function. It includes the part of the
distribution estimated by spline interpolation, and the parts estimated
via the generalized Pareto distribution. This method can only be applied
to objects of class \code{gpinter_dist_orig}, which are returned by
\code{tabulation_fit} and \code{share_fit}.
}
\details{
If \eqn{X} follows the distribution that was interpolated using
\code{tabulation_fit} or \code{share_fit}, then the interpolation function
correspond to:
\deqn{p \mapsto -\log\left(\int_{1 - \exp(-x)}^{1} Q(u) du\right)}
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

