% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-plots.R
\name{plot_tail}
\alias{plot_tail}
\title{Tail function plot}
\usage{
plot_tail(dist, xlim, ...)
}
\arguments{
\item{dist}{A \code{gpinter_dist_orig} object, as returned by
\code{tabulation_fit} or \code{share_fit}.}

\item{xlim}{The range of the curve.}

\item{...}{Ignored.}
}
\description{
Plots the tail function of a distribution estimated via
generalized Pareto interpolation.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

