% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny application in the web browser}
\usage{
run_app()
}
\description{
Run the Shiny app of wid.world/gpinter locally.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

