% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shares-fit.R
\name{shares_fit}
\alias{shares_fit}
\title{Fit a nonparametric distribution on fractile shares}
\usage{
shares_fit(p, average, bracketshare = NULL, topshare = NULL,
  bracketavg = NULL, topavg = NULL, ci = TRUE, samplesize = NULL,
  deriv3max = NULL)
}
\arguments{
\item{p}{A vector of values in [0, 1].}

\item{average}{The average over the entire distribution.}

\item{bracketshare}{The corresponding bracket share.}

\item{topshare}{The corresponding top share.}

\item{bracketavg}{The corresponding bracket average.}

\item{topavg}{The corresponding top average.}

\item{ci}{Should confidence interval (identification region) be computed?
Default is \code{TRUE}.}

\item{samplesize}{The size of the underlying sample.}

\item{deriv3max}{A functional bound of the third derivative of the
interpolation function.}
}
\description{
Estimate a distribution nonparametrically based on tabulations
of a distribution which include:
\itemize{
    \item fractiles of the data,
    \item the share of each bracket
}
}
\author{
Thomas Blanchet
}

