% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-simulation.R
\name{simulate_gpinter}
\alias{simulate_gpinter}
\title{Simulate data}
\usage{
simulate_gpinter(dist, n, ...)
}
\arguments{
\item{dist}{An object of class \code{gpinter_dist_orig}, \code{gpinter_dist_indiv},
\code{gpinter_dist_addup} or \code{gpinter_dist_merge}.}

\item{n}{The size of the sample to generate.}

\item{...}{Ignored.}
}
\value{
A sample of size \code{n}.
}
\description{
Generate a sample from a distribution estimated via
generalized Pareto interpolation.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

