% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline-error-degree2.R
\name{spline_error_degree2}
\alias{spline_error_degree2}
\title{Spline interpolation error over the entire domain with
unknown derivatives}
\usage{
spline_error_degree2(x, xk, upperbound)
}
\arguments{
\item{x}{A value between \code{min(xk)} and \code{max(xk)}.}

\item{xk}{The grid points for the spline estimation.}

\item{upperbound}{A function that majorates the fourth derivative.}
}
\value{
The value of the bound at x.
}
\description{
Gives the bound on the spline interpolation error
for any point of the interpolation domain.
}
\author{
Thomas Blanchet
}

