% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-functions.R
\name{support}
\alias{support}
\title{Support of a distribution estimated via generalized Pareto interpolation}
\usage{
support(dist, ...)
}
\arguments{
\item{dist}{An object of class \code{gpinter_dist_orig}, \code{gpinter_dist_indiv},
\code{gpinter_dist_addup} or \code{gpinter_dist_merge}.}

\item{...}{Ignored.}
}
\value{
A list with two components:
\describe{
    \item{lower}{The lower bound of the distribution (possibly \code{-Inf}).}
    \item{upper}{The upper bound of the distribution (possibly \code{+Inf}).}
}
}
\description{
Give the support of a distribution estimated via generalized
Pareto distribution.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

