% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulation-fit.R
\name{tabulation_fit}
\alias{tabulation_fit}
\title{Fit a nonparametric distribution on tabulated data}
\usage{
tabulation_fit(p, threshold, average, bracketshare = NULL, topshare = NULL,
  bracketavg = NULL, topavg = NULL, invpareto = NULL)
}
\arguments{
\item{p}{A vector of values in [0, 1].}

\item{threshold}{The quantiles corresponding to \code{p}.}

\item{average}{The average over the entire distribution.}

\item{bracketshare}{The corresponding bracket share.}

\item{topshare}{The corresponding top share.}

\item{bracketavg}{The corresponding bracket average.}

\item{topavg}{The corresponding top average.}

\item{invpareto}{The inverted Pareto coefficient.}
}
\value{
An object of class \code{gpinter_dist_orig}.
}
\description{
Estimate a distribution nonparametrically based on tabulations
of a distribution which include:
\itemize{
    \item fractiles of the data,
    \item the corresponding brackets,
    \item the share of each bracket.
}
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

