% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quintic-spline.R
\name{tension}
\alias{tension}
\title{Tension of several splines}
\usage{
tension(xk, yk, sk, ak)
}
\arguments{
\item{xk}{A vector of interpolation points.}

\item{yk}{A vector of values at each interpolation point.}

\item{sk}{A vector of slopes at each interpolation point.}

\item{ak}{A vector of second derivatives at each interpolation point.}
}
\value{
The tension of the spline over [\code{min(xk)}, \code{max(xk)}].
}
\description{
Tension (as mesured by the integral of the square of the
third derivative) of several splines put together.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

