% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quintic-spline.R
\name{tension_grad}
\alias{tension_grad}
\title{Gradient of the tension of several splines}
\usage{
tension_grad(xk, yk, sk, ak)
}
\arguments{
\item{xk}{A vector of interpolation points.}

\item{yk}{A vector of values at each interpolation point.}

\item{sk}{A vector of slopes at each interpolation point.}

\item{ak}{A vector of second derivatives at each interpolation point.}
}
\value{
A vector of length \code{length(ak)} with the derivatives.
}
\description{
The gradient of the tension of several splines with respect
to the value of the second derivative at each knot.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}
