% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{tension_spline}
\alias{tension_spline}
\alias{tension_spline_deriv_a0}
\alias{tension_spline_deriv_a1}
\title{Tension of a quintic spline}
\usage{
tension_spline(x0, x1, y0, y1, s0, s1, a0, a1)
}
\arguments{
\item{x0}{Lower bound of the interval.}

\item{x1}{Upper bound of the interval.}

\item{y0}{Value of the spline at \code{x0} for the first spline.}

\item{y1}{Value of the spline at \code{x1} for the first spline.}

\item{s0}{Slope of the spline at \code{x0} for the first spline.}

\item{s1}{Slope of the spline at \code{x1} for the first spline.}

\item{a0}{Second derivative of the spline at \code{x0} for the first spline.}

\item{a1}{Second derivative of the spline at \code{x1} for the first spline.}
}
\value{
The value of the distance or its derivatives.
}
\description{
Calculates the tension of a quintic spline, as measured by
the square of the integral of the third derivative. Also calculates the
derivative of this tension with respect to the second derivative of the
spline at each bound.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

