% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-functions.R
\name{threshold_share}
\alias{threshold_share}
\title{Share above a threshold}
\usage{
threshold_share(dist, q, ...)
}
\arguments{
\item{dist}{An object of class \code{gpinter_dist_orig}, \code{gpinter_dist_indiv},
\code{gpinter_dist_addup} or \code{gpinter_dist_merge}.}

\item{q}{A vector of real numbers.}

\item{...}{Ignored.}
}
\description{
Calculate the share a threshold. It is mathematically equivalent
to \code{top_share(dist, fitted_cdf(dist, q))}, but can be much faster.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}
