% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-functions.R
\name{top_share}
\alias{top_share}
\title{Top share for generalized Pareto interpolation}
\usage{
top_share(dist, p, ...)
}
\arguments{
\item{dist}{An object of class \code{gpinter_dist_orig}, \code{gpinter_dist_indiv},
\code{gpinter_dist_addup} or \code{gpinter_dist_merge}.}

\item{p}{A vector of probabilities in [0, 1].}

\item{...}{Ignored.}
}
\value{
The value of the share of the top 100*(1 - p)%.
}
\description{
Compute the value of the top share for a distribution estimated
via generalized Pareto interpolation.
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}

