% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulation-fit.R
\name{fitted_density}
\alias{fitted_density}
\title{Density function}
\usage{
fitted_density(q, pk, qk, xk, yk, sk, ak, mu_top, sigma_top, xi_top, mu_bottom,
  sigma_bottom, xi_bottom)
}
\arguments{
\item{q}{The function evaluation point.}

\item{pk}{The probabilities of the tabulation.}

\item{qk}{The associated thresholds of the tabulation.}

\item{xk}{A vector of interpolation points.}

\item{yk}{A vector of values at each interpolation point.}

\item{sk}{A vector of slopes at each interpolation point.}

\item{ak}{A vector of second derivatives at each interpolation point.}

\item{mu_top}{The location parameter for the generalized Pareto distribution at
the top of the distribution.}

\item{sigma_top}{The scale parameter for the generalized Pareto distribution at
the top of the distribution.}

\item{xi_top}{The shape parameter for the generalized Pareto distribution at
the top of the distribution.}

\item{mu_bottom}{The location parameter for the generalized Pareto distribution at
the bottom of the distribution.}

\item{sigma_bottom}{The scale parameter for the generalized Pareto distribution at
the bottom of the distribution.}

\item{xi_bottom}{The shape parameter for the generalized Pareto distribution at
the bottom of the distribution.}
}
\value{
The value of the density at \code{q}.
}
\description{
Full density function, combining the quintic spline
in the middle and the generalized Pareto distribution at the extremes.
}
\author{
Thomas Blanchet
}

