% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation-functions.R
\name{phi}
\alias{phi}
\title{Interpolation function}
\usage{
phi(x, xk, yk, sk, ak, mu_top, sigma_top, xi_top, mu_bottom, sigma_bottom,
  xi_bottom)
}
\arguments{
\item{x}{The function evaluation point.}

\item{xk}{A vector of interpolation points.}

\item{yk}{A vector of values at each interpolation point.}

\item{sk}{A vector of slopes at each interpolation point.}

\item{ak}{A vector of second derivatives at each interpolation point.}

\item{mu_top}{The location parameter for the generalized Pareto distribution at
the top of the distribution.}

\item{sigma_top}{The scale parameter for the generalized Pareto distribution at
the top of the distribution.}

\item{xi_top}{The shape parameter for the generalized Pareto distribution at
the top of the distribution.}

\item{mu_bottom}{The location parameter for the generalized Pareto distribution at
the bottom of the distribution.}

\item{sigma_bottom}{The scale parameter for the generalized Pareto distribution at
the bottom of the distribution.}

\item{xi_bottom}{The shape parameter for the generalized Pareto distribution at
the bottom of the distribution.}
}
\value{
The value of the interpolation function at \code{x}.
}
\description{
Full interpolation function, combining the quintic spline
at the bottom and the generalized Pareto distribution at the top.
}
\author{
Thomas Blanchet
}

