% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shares-fit.R
\name{shares_fit}
\alias{shares_fit}
\title{Fit a nonparametric distribution on tabulated data without thresholds}
\usage{
shares_fit(p, average = NULL, bracketshare = NULL, topshare = NULL,
  bracketavg = NULL, topavg = NULL, first_threshold = NULL,
  bottom_model = NULL, lower_bound = 0)
}
\arguments{
\item{p}{A vector of values in [0, 1].}

\item{average}{The average over the entire distribution.}

\item{bracketshare}{The corresponding bracket share.}

\item{topshare}{The corresponding top share.}

\item{bracketavg}{The corresponding bracket average.}

\item{topavg}{The corresponding top average.}

\item{first_threshold}{The value of the first threshold. If \code{NULL}, it
is estimated from the data. Default is \code{NULL}.}

\item{bottom_model}{Which model to use at the bottom of the distribution?
Only relevant if \code{min(p) > 0}. Either \code{"gpd"} for the generalized
Pareto distribution, or \code{"hist"} for histogram density. Default is
\code{"hist"} if \code{min(threshold) > 0}, and \code{"gpd"} otherwise.}

\item{lower_bound}{Lower bound of the distribution. Only relevant if
\code{min(p) > 0}. Default is \code{0}.}
}
\value{
An object of class \code{gpinter_dist_orig}.
}
\description{
Estimate a distribution nonparametrically based on tabulations
of a distribution which include:
\itemize{
    \item fractiles of the data,
    \item the share of each bracket.
}
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}
