% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulation-fit.R
\name{tabulation_fit}
\alias{tabulation_fit}
\title{Fit a nonparametric distribution on tabulated data}
\usage{
tabulation_fit(p, threshold, average, bracketshare = NULL, topshare = NULL,
  bracketavg = NULL, topavg = NULL, citype = 1, cilevel = 0.05,
  samplesize = NULL, deriv4max = NULL, nrep = 1000)
}
\arguments{
\item{p}{A vector of values in [0, 1].}

\item{threshold}{The quantiles corresponding to \code{p}.}

\item{average}{The average over the entire distribution.}

\item{bracketshare}{The corresponding bracket share.}

\item{topshare}{The corresponding top share.}

\item{bracketavg}{The corresponding bracket average.}

\item{topavg}{The corresponding top average.}

\item{citype}{The type of confidence interval to compute. 0 = no confidence
interval, 1 = spline interpolation error, 2 = spline interpolation error and
sampling error (may be slow).}

\item{cilevel}{The confidence level. Default is 0.05.}

\item{samplesize}{The size of the underlying sample.}

\item{deriv4max}{A function majorant of the fourth derivative of the
interpolation function.}

\item{nrep}{The number of resampling to perform in order to estimate the
sampling error.}
}
\description{
Estimate a distribution nonparametrically based on tabulations
of a distribution which include:
\itemize{
    \item fractiles of the data,
    \item the corresponding brackets,
    \item the share of each bracket.
}
}
\author{
Thomas Blanchet
}

