% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholds-fit.R
\name{thresholds_fit}
\alias{thresholds_fit}
\title{Fit a nonparametric distribution on tabulated data with thresholds only}
\usage{
thresholds_fit(p, threshold, average = NULL, last_bracketshare = NULL,
  last_bracketavg = NULL, last_invpareto = NULL, bottom_model = NULL,
  lower_bound = 0, binf = NULL, fast = FALSE)
}
\arguments{
\item{p}{A vector of values in [0, 1].}

\item{threshold}{The quantiles corresponding to \code{p}.}

\item{average}{The average over the entire distribution. Use \code{NULL} for
unknown. (Default is \code{NULL}.)}

\item{last_bracketshare}{The share of the last bracket. Use \code{NULL} for
unknown. (Default is \code{NULL}.)}

\item{last_bracketavg}{The average in the last bracket. Use \code{NULL} for
unknown. (Default is \code{NULL}.)}

\item{last_invpareto}{The inverted Pareto coefficient at the last threshold.
Use \code{NULL} for unknown. (Default is \code{NULL}.)}

\item{bottom_model}{Which model to use at the bottom of the distribution?
Only relevant if \code{min(p) >= 0}. Either \code{"gpd"} for the generalized
Pareto distribution, \code{"hist"} for histogram density, or \code{"dirac"}.
Default is \code{"hist"} if \code{min(threshold) > 0}, \code{"dirac"} if
\code{min(threshold) == 0} and \code{"gpd"} otherwise.}

\item{lower_bound}{Lower bound of the distribution. Only relevant if
\code{min(p) > 0}. Default is \code{0}.}

\item{binf}{The asymptotic value of the inverted Pareto coefficient. If
\code{NULL}, it is estimated from the data (recommended, unless the
estimated value implies infinite mean). Default is \code{NULL}.}

\item{fast}{Use a faster but less precise method (split-histogram)?
Default is \code{FALSE}.}
}
\value{
An object of class \code{gpinter_dist_orig}.
}
\description{
Estimate a distribution nonparametrically based on tabulations
of a distribution which include:
\itemize{
    \item fractiles of the data,
    \item the corresponding thresholds.
}
}
\author{
Thomas Blanchet, Juliette Fournier, Thomas Piketty
}
